Implemented error if existing contract exists on file

+++ server/database/procedures/document_procedures.sql
@@ -57,5 +57,8 @@ END;
  */
 CREATE PROCEDURE delete_document(IN path_in VARCHAR(500))
 BEGIN
+    IF (SELECT COUNT(*) FROM document d LEFT JOIN contract c on d.document_id = c.document WHERE d.path=path_in) THEN
+      CALL raise(2000, 'Document cannot be deleted because of an existing contract bound to this.');
+    END IF;
     DELETE FROM document WHERE path = path_in;
 END;
